cleanup up parsing code & removed now unnecessary methods
